+++ src/components/Projects/Projects.js
@@ -2,6 +2,7 @@ import { Box, Typography } from '@mui/material';
 import AsyncHelmet from '../AsyncHelmet';
 import projectData from '../../data/projects.json';
 import FilteringMenu from './FilteringMenu';
+import ProjectsContainer from './ProjectsContainer';
 
 const Projects = () => {
   return (
@@ -20,7 +21,7 @@ const Projects = () => {
         Projects
       </Typography>
 
-      <FilteringMenu projectData={projectData} />
+      <ProjectsContainer projectData={projectData} />
     </Box>
   );
 };
